Issue-52: Renamed MainScreenFragment.kt -> PaginatedScreenFragment.kt and MainScreenVM.kt -> PaginatedScreenVM.kt
