+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/movies/MoviesVM.kt
@@ -2,9 +2,9 @@ package com.kirchhoff.movies.ui.screens.main.movies
 
 import com.kirchhoff.movies.data.responses.DiscoverMoviesResponse
 import com.kirchhoff.movies.repository.discover.IDiscoverRepository
-import com.kirchhoff.movies.ui.screens.main.MainScreenVM
+import com.kirchhoff.movies.ui.screens.PaginatedScreenVM
 
 class MoviesVM(private val discoverRepository: IDiscoverRepository) :
-    MainScreenVM<DiscoverMoviesResponse>() {
+    PaginatedScreenVM<DiscoverMoviesResponse>() {
     override suspend fun loadData(page: Int) = discoverRepository.fetchMovies(page)
 }
