+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/persons/PersonsFragment.kt
@@ -3,12 +3,12 @@ package com.kirchhoff.movies.ui.screens.main.persons
 import com.kirchhoff.movies.data.Person
 import com.kirchhoff.movies.data.responses.PersonsResponse
 import com.kirchhoff.movies.ui.screens.details.DetailsActivity
-import com.kirchhoff.movies.ui.screens.main.MainScreenFragment
+import com.kirchhoff.movies.ui.screens.PaginatedScreenFragment
 import com.kirchhoff.movies.ui.screens.main.persons.adapter.PersonsListAdapter
 import com.kirchhoff.movies.ui.utils.recyclerView.BaseRecyclerViewAdapter
 import org.koin.android.viewmodel.ext.android.viewModel
 
-class PersonsFragment : MainScreenFragment<Person, PersonsResponse>(),
+class PersonsFragment : PaginatedScreenFragment<Person, PersonsResponse>(),
     BaseRecyclerViewAdapter.OnItemClickListener<Person> {
 
     override val vm by viewModel<PersonsVM>()
