+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/tvs/TvsFragment.kt
@@ -3,12 +3,12 @@ package com.kirchhoff.movies.ui.screens.main.tvs
 import com.kirchhoff.movies.data.Tv
 import com.kirchhoff.movies.data.responses.DiscoverTvsResponse
 import com.kirchhoff.movies.ui.screens.details.DetailsActivity
-import com.kirchhoff.movies.ui.screens.main.MainScreenFragment
+import com.kirchhoff.movies.ui.screens.PaginatedScreenFragment
 import com.kirchhoff.movies.ui.screens.main.tvs.adapter.TvsListAdapter
 import com.kirchhoff.movies.ui.utils.recyclerView.BaseRecyclerViewAdapter
 import org.koin.android.viewmodel.ext.android.viewModel
 
-class TvsFragment : MainScreenFragment<Tv, DiscoverTvsResponse>(),
+class TvsFragment : PaginatedScreenFragment<Tv, DiscoverTvsResponse>(),
     BaseRecyclerViewAdapter.OnItemClickListener<Tv> {
 
     override val vm by viewModel<TvsVM>()
