+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/tvs/TvsVM.kt
@@ -2,9 +2,9 @@ package com.kirchhoff.movies.ui.screens.main.tvs
 
 import com.kirchhoff.movies.data.responses.DiscoverTvsResponse
 import com.kirchhoff.movies.repository.discover.IDiscoverRepository
-import com.kirchhoff.movies.ui.screens.main.MainScreenVM
+import com.kirchhoff.movies.ui.screens.PaginatedScreenVM
 
 class TvsVM(private val discoverRepository: IDiscoverRepository) :
-    MainScreenVM<DiscoverTvsResponse>() {
+    PaginatedScreenVM<DiscoverTvsResponse>() {
     override suspend fun loadData(page: Int) = discoverRepository.fetchTvs(page)
 }
